Remove old download queue
